Moved GetThisEnvironment to ExecutionContext for reuse in CallFrame.This
