+++ Jint/Runtime/Interpreter/Expressions/JintMemberExpression.cs
@@ -65,7 +65,7 @@ namespace Jint.Runtime.Interpreter.Expressions
             }
             else if (_objectExpression is JintSuperExpression)
             {
-                var env = (FunctionEnvironmentRecord) _engine.GetThisEnvironment();
+                var env = (FunctionEnvironmentRecord) _engine.ExecutionContext.GetThisEnvironment();
                 actualThis = env.GetThisBinding();
                 baseValue = env.GetSuperBase();
             }
